test_setup.py
#!/usr/bin/env python3

print("🚀 Testing LLM Query Retrieval System Setup")
print("=" * 50)

# Test 1: Basic imports
try:
    import os
    import numpy as np
    import pandas as pd
    print("✅ Basic imports: OK")
except Exception as e:
    print(f"❌ Basic imports failed: {e}")

# Test 2: Environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
    github_token = os.getenv("GITHUB_TOKEN")
    if github_token and github_token != "your_github_token_here":
        print("✅ Environment variables: OK")
    else:
        print("⚠️ Environment variables: Missing tokens")
except Exception as e:
    print(f"❌ Environment loading failed: {e}")

# Test 3: Database packages
try:
    import sqlalchemy
    import psycopg2
    print("✅ Database packages: OK")
except Exception as e:
    print(f"❌ Database packages failed: {e}")

# Test 4: Pinecone
try:
    import pinecone
    print("✅ Pinecone: OK")
except Exception as e:
    print(f"❌ Pinecone failed: {e}")

# Test 5: Simple embedding model
try:
    from models.simple_embedding import embedding_model
    info = embedding_model.get_model_info()
    print(f"✅ Simple embedding model: {info}")
except Exception as e:
    print(f"❌ Simple embedding model failed: {e}")

# Test 6: Test encoding
try:
    from models.simple_embedding import embedding_model
    test_text = "What is machine learning?"
    embedding = embedding_model.encode_query(test_text)
    print(f"✅ Text encoding: Shape {embedding.shape}")
except Exception as e:
    print(f"❌ Text encoding failed: {e}")

# Test 7: GitHub Models API
try:
    from openai import OpenAI
    import os
    
    token = os.environ.get("GITHUB_TOKEN")
    if token:
        endpoint = "https://models.github.ai/inference"
        client = OpenAI(base_url=endpoint, api_key=token)
        print("✅ GitHub Models API client: OK")
    else:
        print("⚠️ GitHub Models API: No token found")
except Exception as e:
    print(f"❌ GitHub Models API failed: {e}")

print("\n🎯 Setup Status Summary:")
print("- All core packages are installed")
print("- Simple embedding model is ready")
print("- Database connections can be established")
print("- Ready for data processing and training!")

print("\n📝 Next steps:")
print("1. Add your API keys to .env file")
print("2. Set up PostgreSQL database")
print("3. Configure Pinecone account")
print("4. Run data processing pipeline")
