requirements.txt
pinecone-client==2.2.4
psycopg2-binary==2.9.7
sentence-transformers==2.2.2
langchain==0.0.340
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
datasets==2.14.6
transformers==4.35.2
torch==2.1.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
